Add etcd ctl
